==> python_35_code/5_Simulated_Classification_Dataset_ML_pipeline_py35.ipynb <==
# Libraries

%matplotlib inline
import matplotlib
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns
from __future__ import print_function

# Versions

**Disclaimer: This notebook uses Python 3.5. Sections may not work if you try this with Python 2.X.**

items = [("Numpy", np), ("Pandas", pd), ("Matplotlib", matplotlib), ("Seaborn", sns)]
for item in items:
    print(item[0] + " version " + str(item[1].__version__))

---

# Get Data

X_train = pd.read_hdf('/Users/davidziganto/Repositories/Synthetic_Dataset_Generation/data/py35/simulated_cleaned_training_data_py35.h5', 'table')
y_train = X_train.pop('hired')

# Get X_test & y_test

import pickle

path = "/Users/davidziganto/Repositories/Synthetic_Dataset_Generation/pickle_files/py35/"

with open(path + "X_test_py35.pkl", 'rb') as picklefile: 
    X_test = pickle.load(picklefile)
    
with open(path + "y_test_py35.pkl", 'rb') as picklefile: 
    y_test = pickle.load(picklefile)

# Machine Learning

This section shows the machine learning pipeline used to generate scoring thresholds. A further discussion of how this information is used will be included in the next notebook.

### Check Class Weights

pos_class_prop = y_train.mean()
pos_class_prop

### ML Libraries

from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.dummy import DummyClassifier
from sklearn.metrics import log_loss, confusion_matrix
from sklearn.model_selection import RandomizedSearchCV, GridSearchCV

### Instantiate Models

dt = DecisionTreeClassifier(criterion='gini',
                            splitter='best', 
                            max_depth=None, 
                            min_samples_split=2, 
                            min_samples_leaf=1, 
                            min_weight_fraction_leaf=0.0, 
                            max_features=None, 
                            random_state=13, 
                            max_leaf_nodes=None, 
                            min_impurity_split=1e-07, 
                            class_weight=None, 
                            presort=False)

dummy = DummyClassifier(strategy='most_frequent', 
                        random_state=99, 
                        constant=None)

gbc = GradientBoostingClassifier(loss='deviance', 
                                 learning_rate=0.1, 
                                 n_estimators=50, 
                                 subsample=1.0, 
                                 criterion='friedman_mse', 
                                 min_samples_split=2, 
                                 min_samples_leaf=1, 
                                 min_weight_fraction_leaf=0.0, 
                                 max_depth=3, 
                                 min_impurity_split=1e-07, 
                                 init=None, 
                                 random_state=123, 
                                 max_features=None, 
                                 verbose=0, 
                                 max_leaf_nodes=None, 
                                 warm_start=False, 
                                 presort='auto')

knn = KNeighborsClassifier(n_neighbors=5, 
                           weights='uniform', 
                           algorithm='auto', 
                           leaf_size=30, 
                           p=2, 
                           metric='euclidean', 
                           metric_params=None, 
                           n_jobs=-1)

lr = LogisticRegression(penalty='l2', 
                        dual=False, 
                        tol=0.0001, 
                        C=1.0, 
                        fit_intercept=True, 
                        intercept_scaling=1, 
                        class_weight=None, 
                        random_state=10, 
                        solver='liblinear', 
                        max_iter=100, 
                        multi_class='ovr', 
                        verbose=0, 
                        warm_start=False, 
                        n_jobs=-1)


nb = MultinomialNB(alpha=1.0, 
                   fit_prior=True, 
                   class_prior=None)

rf = RandomForestClassifier(n_estimators=50, 
                            criterion='gini', 
                            max_depth=None, 
                            min_samples_split=2, 
                            min_samples_leaf=1, 
                            min_weight_fraction_leaf=0.0, 
                            max_features='auto', 
                            max_leaf_nodes=None, 
                            min_impurity_split=1e-07, 
                            bootstrap=True, 
                            oob_score=False, 
                            n_jobs=-1, 
                            random_state=17, 
                            verbose=0, 
                            warm_start=False, 
                            class_weight=None)

### Setup Parameter Grid For RandomCV

dt_param_grid = dict(max_depth=[None, 6, 8], 
                     min_samples_leaf=range(1,5),
                     class_weight=[None, 'balanced'])

gbc_param_grid = dict(loss=['deviance','exponential'], 
                      max_depth=range(2,5), 
                      learning_rate=[0.001, 0.01, 0.1])

knn_param_grid = dict(n_neighbors=range(1, 15, 2), 
                      weights=('uniform', 'distance'))

lr_param_grid = dict(penalty=['l1', 'l2'], 
                     C=np.geomspace(0.001, 10, num=5),
                     class_weight=[None, 'balanced'])

rf_param_grid = dict(n_estimators=[50, 500], 
                     max_depth=[None, 8, 10],
                     class_weight=[None, 'balanced'])

### Setup Dictionary For Algo_Report()

from collections import OrderedDict

algo_dict = OrderedDict(
    (
        ("dt",(dt, dt_param_grid)),
        ("dummy",(dummy)),
        ("gbc",(gbc, gbc_param_grid)),
        ("knn",(knn, knn_param_grid)),
        ("lr",(lr, lr_param_grid)), 
        ("nb",(nb)), 
        ("rf",(rf, rf_param_grid))
    )
)

### Algo_Report()

def algo_report(algo_dict, cv=5, search_flag=1):
    '''
    Function that generates in-sample and out-of-sample metrics for numerous machine learning algorithms.
    
    Input:
        algo_dict = dictionary with algorithm name as key and model object & parameter grid as values
        cv = number of folds for cross validation
        search_flag = {0: use default model paramters; 1: use randomized search; 2: use grid search}
    Output:
        prints a report showing:
            1) in-sample negative log-loss value or accuracy (dependent on score function)
            2) out-of-sample negative log-loss value or accuracy (dependent on score function)
            3) out-of-sample log loss value
            4) confusion matrix
    '''
    for k, v in algo_dict.items():  
        if k == "nb" or k == "dummy":
            model = v.fit(X_train, y_train)
        else:
            if search_flag:
                model = RandomizedSearchCV(v[0], v[1], cv=cv, scoring='neg_log_loss', random_state=42)
                model.fit(X_train, y_train)
            elif search_flag==2:
                model = GridSearchCV(v[0], v[1], cv=cv, scoring='neg_log_loss', random_state=42)
                model.fit(X_train, y_train)
            else: 
                model = v[0].fit(X_train, y_train)

        print("[%s]" % k)
        print("In-Sample:     {}\nOut-of_Sample: {}\nLog_loss:      {}".format(
            round(model.score(X_train, y_train),3), 
            round(model.score(X_test, y_test),3), 
            round(log_loss(y_test, model.predict_proba(X_test)),3)))
        print("\nConfusion Matrix:")
        print(confusion_matrix(y_test, model.predict(X_test)))
        print("\n-----------------\n")

### Results
Results for all three scenarios - using only defaults, using random search, and using grid search - can be found below. 

#### Defaults

algo_report(algo_dict, cv=10, search_flag=0)

#### Random Search

algo_report(algo_dict, cv=10, search_flag=1)

#### Grid Search

algo_report(algo_dict, cv=10, search_flag=2)

# Create Models For Pickling

# Needs Improvement Model
knn_randcv = RandomizedSearchCV(knn, knn_param_grid, cv=10, scoring='neg_log_loss', random_state=42)
knn_randcv.fit(X_train, y_train)
log_loss(y_test, knn_randcv.predict_proba(X_test))

# Satisfactory Model
rf_randcv = RandomizedSearchCV(rf, rf_param_grid, cv=10, scoring='neg_log_loss', random_state=42)
rf_randcv.fit(X_train, y_train)
log_loss(y_test, rf_randcv.predict_proba(X_test))

# Proficient Model
gbc_randcv = RandomizedSearchCV(gbc, gbc_param_grid, cv=10, scoring='neg_log_loss', random_state=42)
gbc_randcv.fit(X_train, y_train)
log_loss(y_test, gbc_randcv.predict_proba(X_test))

# Pickle Models & Test Set Data For Auto-Scoring

path = '/Users/davidziganto/Repositories/Synthetic_Dataset_Generation/pickle_files/py35/'

# Save KNN model
with open(path + 'knn_needs_improvement_py35.pkl', 'wb') as picklefile:
    pickle.dump(knn_randcv, picklefile)
    
# Save Random Forest model
with open(path + 'rf_satisfactory_py35.pkl', 'wb') as picklefile:
    pickle.dump(rf_randcv, picklefile)
    
# Save Gradient Boosted Classifier model
with open(path + 'gbc_proficient_py35.pkl', 'wb') as picklefile:
    pickle.dump(gbc_randcv, picklefile)
